Test if buildDir exists before call forceDelete
git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@163010 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-mboot2/src/main/java/MBoot.java
@@ -402,7 +402,10 @@ public class MBoot
 
         // clean
         System.out.println( "Cleaning " + buildDirFile + "..." );
-        FileUtils.forceDelete( buildDirFile );
+        if ( buildDirFile.exists() )
+        {
+            FileUtils.forceDelete( buildDirFile );
+        }
 
         // ----------------------------------------------------------------------
         // Download deps
